Demonstrate transfer between two standalone rabbitmq nodes
Before this commit, we used one node and one target queue.
With this commit, we deploy -via docker- two standalone nodes
representing two separate rabbitmq clusters. The clusters are
named after the diagram illustrated in the README.md. They are
PC1-OKR and LTM-OKR. The messages will be in LTM-OKR and they
will be transfered to PC1-OKR.

Scripts which starts with underscore are not meant to invoked
directly by the user.
Only rabbit/deploy, rabbit/undeploy, transfer and
 setup_vhost_in_both_clusters are meant for end users